+++ NineInTeam/NineInTeam/Presentation/Membership/SignIn/LoginServiceProvider.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-enum LoginServiceProvider: Codable {
+// apple, none.. 등 로그인 서비스에 따른 분기작업시 사용
+enum SignInServiceProvider: Codable {
     case kakao
-    // apple, none.. 등 로그인 서비스에 따른 분기작업시 사용
 }
