+++ NineInTeam/NineInTeam/Util/Error/CustomEncodingError.swift
@@ -1,0 +1,14 @@
+//
+//  CustomEncodingError.swift
+//  NineInTeam
+//
+//  Created by HeonJin Ha on 2023/06/20.
+//
+
+import Foundation
+
+enum CustomEncodingError: Error {
+    
+    case invalidData
+    
+}
